Extract tool surface build helpers into tool_surface/build.rs
Moves the 8 free-function build helpers (downgrade_file_tools,
downgrade_bash, build_host_tools, resolve_effective_tool_root,
resolve_configured_tool_root, resolve_path_with_canonical_prefix,
dedupe_strings, has_registered_mcp_services) from mod.rs into
build.rs with pub(super) visibility. Updates behavior_config.rs
and the mod.rs call site accordingly.
