+++ crates/defra-agent/src/tool_surface/behavior_config.rs
@@ -5,10 +5,11 @@ use crate::toolset::ToolSet;
 
 use super::modes::ToolCeiling;
 use super::selection::{CustomToolFactory, ToolSelection};
-use super::{
+use super::build::{
     build_host_tools, dedupe_strings, downgrade_bash, downgrade_file_tools,
-    has_registered_mcp_services, ToolSurface,
+    has_registered_mcp_services,
 };
+use super::ToolSurface;
 
 #[derive(Clone)]
 pub struct BehaviorToolConfig {
