+++ crates/defra-agent/src/tool_surface/mod.rs
@@ -9,17 +9,14 @@ pub use modes::{BashMode, FileToolMode, ToolCeiling};
 pub use runtime_context::ToolRuntimeContext;
 pub use selection::{CustomToolFactory, ToolSelection};
 
-use std::collections::{HashMap, HashSet};
-use std::path::{Path, PathBuf};
+use std::collections::HashMap;
+use std::path::PathBuf;
 
-use anyhow::{anyhow, bail, Context, Result};
-use defra_node::EmbeddedNode;
+use anyhow::Result;
 use rig::tool::ToolDyn;
 
 use crate::meta_tools::{build_meta_tools, META_TOOL_NAMES};
-use crate::toolset::{
-    build_delegate_tool, CliToolConfig, ToolSet, ToolSetBuilder, DELEGATE_TOOL_NAME,
-};
+use crate::toolset::{build_delegate_tool, CliToolConfig, ToolSet, DELEGATE_TOOL_NAME};
 
 const DEFAULT_CLI_TIMEOUT_SECS: u64 = 10;
 
@@ -53,7 +50,7 @@ impl ToolSurface {
             names.push(DELEGATE_TOOL_NAME.to_string());
         }
         names.extend(self.custom_tools.iter().map(|tool| tool.name().to_string()));
-        dedupe_strings(names)
+        build::dedupe_strings(names)
     }
 
     pub fn build_tools(&self, runtime: &ToolRuntimeContext) -> Result<Vec<Box<dyn ToolDyn>>> {
@@ -114,214 +111,6 @@ pub fn cli_tool(
     }
 }
 
-fn downgrade_file_tools(
-    behavior_name: &str,
-    requested: FileToolMode,
-    ceiling: FileToolMode,
-) -> FileToolMode {
-    if requested.rank() <= ceiling.rank() {
-        return requested;
-    }
-
-    tracing::warn!(
-        behavior_id = %behavior_name,
-        requested = ?requested,
-        ceiling = ?ceiling,
-        "downgrading file tool mode to fit tool ceiling"
-    );
-    ceiling
-}
-
-fn downgrade_bash(behavior_name: &str, requested: BashMode, ceiling: BashMode) -> BashMode {
-    if requested.rank() <= ceiling.rank() {
-        return requested;
-    }
-
-    tracing::warn!(
-        behavior_id = %behavior_name,
-        requested = ?requested,
-        ceiling = ?ceiling,
-        "downgrading bash mode to fit tool ceiling"
-    );
-    ceiling
-}
-
-fn build_host_tools(
-    behavior_name: &str,
-    file_tools: FileToolMode,
-    bash: BashMode,
-    file_tool_root: Option<&Path>,
-    cli_tool_names: &[String],
-    ceiling: &ToolCeiling,
-) -> Result<ToolSet> {
-    let mut builder = ToolSetBuilder::default();
-    let needs_file_tool_root =
-        !matches!(file_tools, FileToolMode::Off) || !matches!(bash, BashMode::Off);
-    let effective_root = if needs_file_tool_root {
-        resolve_effective_tool_root(behavior_name, file_tool_root, ceiling.root())?
-    } else {
-        None
-    };
-    if let Some(root) = effective_root.clone() {
-        builder = builder.read_root(root.clone());
-    }
-
-    if !matches!(file_tools, FileToolMode::Off) {
-        builder = builder.list_files().read_file().glob().grep();
-    }
-
-    if matches!(file_tools, FileToolMode::ReadWrite) {
-        let root = effective_root
-            .clone()
-            .ok_or_else(|| anyhow!("readwrite file tools require a configured tool root"))?;
-        builder = builder.write_file(root.clone()).edit_file(root);
-    }
-
-    match bash {
-        BashMode::Off => {}
-        BashMode::ReadOnly => {
-            builder = builder.bash_read_only();
-        }
-        BashMode::Unrestricted => {
-            let root = effective_root
-                .clone()
-                .ok_or_else(|| anyhow!("unrestricted bash requires a configured tool root"))?;
-            builder = builder.bash_unrestricted(root);
-        }
-    }
-
-    let cli_tools = ceiling
-        .cli_tools()
-        .iter()
-        .map(|tool| (tool.name.clone(), tool.clone()))
-        .collect::<HashMap<_, _>>();
-    for tool_name in dedupe_strings(cli_tool_names.to_vec()) {
-        match cli_tools.get(&tool_name) {
-            Some(tool) => builder = builder.cli_tool(tool.clone()),
-            None => tracing::warn!(
-                behavior_id = %behavior_name,
-                cli_tool = %tool_name,
-                "dropping CLI tool not present in tool ceiling"
-            ),
-        }
-    }
-
-    Ok(builder.build())
-}
-
-fn resolve_effective_tool_root(
-    behavior_name: &str,
-    selection_root: Option<&Path>,
-    ceiling_root: Option<&Path>,
-) -> Result<Option<PathBuf>> {
-    let selection_root = selection_root
-        .map(resolve_configured_tool_root)
-        .transpose()?;
-    let ceiling_root = ceiling_root.map(resolve_configured_tool_root).transpose()?;
-
-    match (selection_root, ceiling_root) {
-        (Some(selection_root), Some(ceiling_root)) => {
-            if selection_root.starts_with(&ceiling_root) {
-                Ok(Some(selection_root))
-            } else {
-                bail!(
-                    "behavior {behavior_name} file tool root {} escapes operator tool root {}",
-                    selection_root.display(),
-                    ceiling_root.display()
-                );
-            }
-        }
-        (Some(selection_root), None) => Ok(Some(selection_root)),
-        (None, Some(ceiling_root)) => Ok(Some(ceiling_root)),
-        (None, None) => Ok(None),
-    }
-}
-
-fn resolve_configured_tool_root(path: &Path) -> Result<PathBuf> {
-    let absolute = if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        std::env::current_dir()
-            .with_context(|| format!("resolving relative tool root {}", path.display()))?
-            .join(path)
-    };
-
-    resolve_path_with_canonical_prefix(&absolute)
-}
-
-fn resolve_path_with_canonical_prefix(path: &Path) -> Result<PathBuf> {
-    let mut resolved = PathBuf::new();
-    let mut missing_tail = false;
-
-    for component in path.components() {
-        match component {
-            std::path::Component::Prefix(_) | std::path::Component::RootDir => {
-                resolved.push(component.as_os_str());
-            }
-            std::path::Component::CurDir => {}
-            std::path::Component::ParentDir => {
-                resolved.pop();
-            }
-            std::path::Component::Normal(name) => {
-                let candidate = resolved.join(name);
-                if !missing_tail && candidate.exists() {
-                    resolved = std::fs::canonicalize(&candidate).with_context(|| {
-                        format!("canonicalizing tool root {}", candidate.display())
-                    })?;
-                } else {
-                    missing_tail = true;
-                    resolved.push(name);
-                }
-            }
-        }
-    }
-
-    Ok(resolved)
-}
-
-fn dedupe_strings(values: Vec<String>) -> Vec<String> {
-    let mut seen = HashSet::new();
-    let mut deduped = Vec::with_capacity(values.len());
-    for value in values {
-        let trimmed = value.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        if seen.insert(trimmed.to_string()) {
-            deduped.push(trimmed.to_string());
-        }
-    }
-    deduped
-}
-
-async fn has_registered_mcp_services(node: &EmbeddedNode) -> Result<bool> {
-    let query = r#"{
-  ToolServiceRegistry(
-    filter: { status: { _eq: "online" } }
-    limit: 1
-  ) {
-    service_id
-  }
-}"#;
-
-    let response = node.execute(query).await;
-    if response.has_errors() {
-        bail!(
-            "query ToolServiceRegistry for tool-surface resolution failed: {:?}",
-            response.errors
-        );
-    }
-
-    let services = response
-        .data
-        .as_ref()
-        .and_then(|data| data.get("ToolServiceRegistry"))
-        .and_then(|value| value.as_array())
-        .cloned()
-        .unwrap_or_default();
-
-    Ok(!services.is_empty())
-}
 
 #[cfg(test)]
 mod tests {
